[chore/#142] Component 폴더 삭제, 파일 분리 이동
