+++ ILSANG/Sources/Global/Model/Icon.swift
@@ -2,30 +2,11 @@
 //  Icon.swift
 //  ILSANG
 //
-//  Created by Lee Jinhee on 5/23/24.
+//  Created by Kim Andrew on 9/11/24.
 //
 
+import Foundation
 import SwiftUI
-
-struct IconView: View {
-    let iconWidth: CGFloat
-    let size: IconSize
-    let icon: ImageResource
-    let color: IconColor
-    
-    var body: some View {
-        Image(icon)
-            .resizable()
-            .scaledToFit()
-            .frame(width: iconWidth)
-            .foregroundStyle(color.fgColor)
-            .frame(width: size.value, height: size.value)
-            .background(
-                Circle()
-                    .foregroundStyle(color.bgColor)
-            )
-    }
-}
 
 enum IconSize {
     case small
